Cleaning up SMPI stuff

+++ include/wrench/communicator/Communicator.h
@@ -35,7 +35,7 @@ namespace wrench {
         std::map<aid_t, unsigned long> actor_to_rank;
         std::map<unsigned long, simgrid::s4u::Mailbox *> rank_to_mailbox;
 
-        static void performAllToAll(const std::vector<simgrid::s4u::Host *> &hosts, double bytes) ;
+        static void performAllToAll(const std::vector<simgrid::s4u::Host *> &hosts, int bytes) ;
 
     };
